+++ MaxwellRenderToolbox/docs/version_history.md
@@ -1,6 +1,11 @@
 <a name="version-history"></a>
 # Version History #
 
+## Version 0.1.1 - 2015-12-11 ##
+
+- Added the MXI to GearVR Cubic Panorama Converter script "mxi2gearvrcube.py"
+- Added the MXI to Google Photosphere panoramic metadata embedding script "mxi2photosphere.py"
+
 ## Version 0.1 - 2015-12-08 ##
 
 - Inital GitHub commit